# api/openapi.yaml
openapi: 3.0.3
info:
  title: Inspection Platform API
  version: 0.1.0
  description: 设备巡检与维护平台 REST API。
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /api/auth/login:
    post:
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        "200":
          description: 登录成功，返回短期访问令牌。
  /api/devices:
    get:
      summary: 筛选设备台账。
      parameters:
        - in: query
          name: query
          schema: { type: string }
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: owner
          schema: { type: string }
        - in: query
          name: location
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: 分页设备列表。
    post:
      summary: 创建设备档案。
      responses:
        "201":
          description: 创建成功。
  /api/devices/{id}/history:
    get:
      summary: 获取设备巡检和维修历史。
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: 设备历史。
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT